requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Audio processing
openai-whisper==20231117
faster-whisper==1.0.0
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2

# Audio utilities
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1
webrtcvad==2.0.10

# LLM integration
openai==1.12.0
anthropic==0.18.1

# Document generation
python-docx==1.1.0
reportlab==4.0.9
Pillow==10.2.0

# Data storage
sqlalchemy==2.0.25

# Utilities
pydantic==2.6.1
jsonschema==4.21.1
python-dateutil==2.8.2
